refactor: Move UI functions to utils.py and rename to prompt_*
