+++ src/heretic/model.py
@@ -57,6 +57,18 @@ class Model:
             if hasattr(self.tokenizer, "padding_side"):
                 self.tokenizer.padding_side = "left"
 
+        # Fallback for tokenizers that don't have a chat template (e.g. base models).
+        # We use ChatML as a reasonable default.
+        if self.tokenizer.chat_template is None:
+            self.tokenizer.chat_template = (
+                "{% for message in messages %}"
+                "{{'<|im_start|>' + message['role'] + '\n' + message['content'] + '<|im_end|>' + '\n'}}"
+                "{% endfor %}"
+                "{% if add_generation_prompt %}"
+                "{{ '<|im_start|>assistant\n' }}"
+                "{% endif %}"
+            )
+
         self.model = None
 
         for dtype in settings.dtypes:
